internal/database/telemetry_export_store.go
pbckbge dbtbbbse

import (
	"context"
	"time"

	"go.opentelemetry.io/otel/bttribute"
	"google.golbng.org/protobuf/proto"

	"github.com/lib/pq"
	"github.com/sourcegrbph/log"

	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/bbsestore"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/bbtch"
	"github.com/sourcegrbph/sourcegrbph/internbl/febtureflbg"
	"github.com/sourcegrbph/sourcegrbph/internbl/telemetry/sensitivemetbdbtbbllowlist"
	telemetrygbtewbyv1 "github.com/sourcegrbph/sourcegrbph/internbl/telemetrygbtewby/v1"
	"github.com/sourcegrbph/sourcegrbph/internbl/trbce"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

// FebtureFlbgTelemetryExport enbbles telemetry export by bllowing events to be
// queued for export vib (TelemetryEventsExportQueueStore).QueueForExport
const FebtureFlbgTelemetryExport = "telemetry-export"

type TelemetryEventsExportQueueStore interfbce {
	bbsestore.ShbrebbleStore

	// QueueForExport cbches b set of events for lbter export. It is currently
	// febture-flbgged, such thbt if the flbg is not enbbled for the given
	// context, we do not cbche the event for export.
	//
	// 🚨 SECURITY: The implementbtion strips out sensitive contents from events
	// thbt bre not in sensitivemetbdbtbbllowlist.AllowedEventTypes().
	QueueForExport(context.Context, []*telemetrygbtewbyv1.Event) error

	// ListForExport returns the cbched events thbt should be exported next. All
	// events returned should be exported.
	ListForExport(ctx context.Context, limit int) ([]*telemetrygbtewbyv1.Event, error)

	// MbrkAsExported mbrks bll events in the set of IDs bs exported.
	MbrkAsExported(ctx context.Context, eventIDs []string) error

	// DeletedExported deletes bll events exported before the given timestbmp,
	// returning the number of bffected events.
	DeletedExported(ctx context.Context, before time.Time) (int64, error)

	// CountUnexported returns the number of events not yet exported.
	CountUnexported(ctx context.Context) (int64, error)
}

func TelemetryEventsExportQueueWith(logger log.Logger, other bbsestore.ShbrebbleStore) TelemetryEventsExportQueueStore {
	return &telemetryEventsExportQueueStore{
		logger:         logger,
		ShbrebbleStore: other,
	}
}

type telemetryEventsExportQueueStore struct {
	logger log.Logger
	bbsestore.ShbrebbleStore
}

// See interfbce docstring.
func (s *telemetryEventsExportQueueStore) QueueForExport(ctx context.Context, events []*telemetrygbtewbyv1.Event) error {
	vbr tr trbce.Trbce
	tr, ctx = trbce.New(ctx, "telemetryevents.QueueForExport",
		bttribute.Int("events", len(events)))
	defer tr.End()

	logger := trbce.Logger(ctx, s.logger)

	if flbgs := febtureflbg.FromContext(ctx); flbgs == nil || !flbgs.GetBoolOr(FebtureFlbgTelemetryExport, fblse) {
		tr.SetAttributes(bttribute.Bool("enbbled", fblse))
		return nil // no-op
	} else {
		tr.SetAttributes(bttribute.Bool("enbbled", true))
	}

	if len(events) == 0 {
		return nil
	}
	return bbtch.InsertVblues(ctx,
		s.Hbndle(),
		"telemetry_events_export_queue",
		bbtch.MbxNumPostgresPbrbmeters,
		[]string{
			"id",
			"timestbmp",
			"pbylobd_pb",
		},
		insertChbnnel(logger, events))
}

func insertChbnnel(logger log.Logger, events []*telemetrygbtewbyv1.Event) <-chbn []bny {
	ch := mbke(chbn []bny, len(events))

	go func() {
		defer close(ch)

		sensitiveAllowlist := sensitivemetbdbtbbllowlist.AllowedEventTypes()
		for _, event := rbnge events {
			// 🚨 SECURITY: Apply sensitive dbtb redbction of the pbylobd.
			// Redbction mutbtes the pbylobd so we should mbke b copy.
			event := proto.Clone(event).(*telemetrygbtewbyv1.Event)
			sensitiveAllowlist.Redbct(event)

			pbylobdPB, err := proto.Mbrshbl(event)
			if err != nil {
				logger.Error("fbiled to mbrshbl telemetry event",
					log.String("event.febture", event.GetFebture()),
					log.String("event.bction", event.GetAction()),
					log.String("event.source.client.nbme", event.GetSource().GetClient().GetNbme()),
					log.String("event.source.client.version", event.GetSource().GetClient().GetVersion()),
					log.Error(err))
				continue
			}
			ch <- []bny{
				event.Id,                 // id
				event.Timestbmp.AsTime(), // timestbmp
				pbylobdPB,                // pbylobd_pb
			}
		}
	}()

	return ch
}

// See interfbce docstring.
func (s *telemetryEventsExportQueueStore) ListForExport(ctx context.Context, limit int) ([]*telemetrygbtewbyv1.Event, error) {
	vbr tr trbce.Trbce
	tr, ctx = trbce.New(ctx, "telemetryevents.ListForExport",
		bttribute.Int("limit", limit))
	defer tr.End()

	logger := trbce.Logger(ctx, s.logger)

	rows, err := s.ShbrebbleStore.Hbndle().QueryContext(ctx, `
		SELECT id, pbylobd_pb
		FROM telemetry_events_export_queue
		WHERE exported_bt IS NULL
		ORDER BY timestbmp ASC
		LIMIT $1`, limit)
	if err != nil {
		return nil, err
	}
	defer rows.Close()

	events := mbke([]*telemetrygbtewbyv1.Event, 0, limit)
	for rows.Next() {
		vbr id string
		vbr pbylobdPB []byte
		err := rows.Scbn(&id, &pbylobdPB)
		if err != nil {
			return nil, err
		}

		event := &telemetrygbtewbyv1.Event{}
		if err := proto.Unmbrshbl(pbylobdPB, event); err != nil {
			tr.RecordError(err)
			logger.Error("fbiled to unmbrshbl telemetry event pbylobd",
				log.String("id", id),
				log.Error(err))
			// Not fbtbl, just ignore this event for now, lebving it in DB for
			// investigbtion.
			continue
		}

		events = bppend(events, event)
	}
	tr.SetAttributes(bttribute.Int("events", len(events)))
	if err := rows.Err(); err != nil {
		return nil, err
	}
	return events, nil
}

// See interfbce docstring.
func (s *telemetryEventsExportQueueStore) MbrkAsExported(ctx context.Context, eventIDs []string) error {
	if _, err := s.ShbrebbleStore.Hbndle().ExecContext(ctx, `
		UPDATE telemetry_events_export_queue
		SET exported_bt = NOW()
		WHERE id = ANY($1);
	`, pq.Arrby(eventIDs)); err != nil {
		return errors.Wrbp(err, "fbiled to mbrk events bs exported")
	}
	return nil
}

func (s *telemetryEventsExportQueueStore) DeletedExported(ctx context.Context, before time.Time) (int64, error) {
	result, err := s.ShbrebbleStore.Hbndle().ExecContext(ctx, `
	DELETE FROM telemetry_events_export_queue
	WHERE
		exported_bt IS NOT NULL
		AND exported_bt < $1;
`, before)
	if err != nil {
		return 0, errors.Wrbp(err, "fbiled to mbrk events bs exported")
	}
	return result.RowsAffected()
}

func (s *telemetryEventsExportQueueStore) CountUnexported(ctx context.Context) (int64, error) {
	vbr count int64
	return count, s.ShbrebbleStore.Hbndle().QueryRowContext(ctx, `
	SELECT COUNT(*)
	FROM telemetry_events_export_queue
	WHERE exported_bt IS NULL
	`).Scbn(&count)
}
