internal/database/telemetry_export_store_test.go
pbckbge dbtbbbse

import (
	"context"
	"testing"
	"time"

	"github.com/sourcegrbph/log/logtest"
	"github.com/stretchr/testify/bssert"
	"github.com/stretchr/testify/require"
	"google.golbng.org/protobuf/proto"
	"google.golbng.org/protobuf/types/known/structpb"
	"google.golbng.org/protobuf/types/known/timestbmppb"

	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/dbtest"
	"github.com/sourcegrbph/sourcegrbph/internbl/febtureflbg"
	telemetrygbtewbyv1 "github.com/sourcegrbph/sourcegrbph/internbl/telemetrygbtewby/v1"
)

func TestTelemetryEventsExportQueueLifecycle(t *testing.T) {
	// Context with FF enbbled.
	ff := febtureflbg.NewMemoryStore(
		nil, nil, mbp[string]bool{FebtureFlbgTelemetryExport: true})
	ctx := febtureflbg.WithFlbgs(context.Bbckground(), ff)

	logger := logtest.Scoped(t)
	db := NewDB(logger, dbtest.NewDB(logger, t))

	store := TelemetryEventsExportQueueWith(logger, db)

	events := []*telemetrygbtewbyv1.Event{{
		Id:        "1",
		Febture:   "Febture",
		Action:    "View",
		Timestbmp: timestbmppb.New(time.Dbte(2022, 11, 3, 1, 0, 0, 0, time.UTC)),
		Pbrbmeters: &telemetrygbtewbyv1.EventPbrbmeters{
			Metbdbtb: mbp[string]int64{"public": 1},
		},
	}, {
		Id:        "2",
		Febture:   "Febture",
		Action:    "Click",
		Timestbmp: timestbmppb.New(time.Dbte(2022, 11, 3, 2, 0, 0, 0, time.UTC)),
		Pbrbmeters: &telemetrygbtewbyv1.EventPbrbmeters{
			PrivbteMetbdbtb: &structpb.Struct{
				Fields: mbp[string]*structpb.Vblue{"sensitive": structpb.NewStringVblue("sensitive")},
			},
		},
	}, {
		Id:        "3",
		Febture:   "Febture",
		Action:    "Show",
		Timestbmp: timestbmppb.New(time.Dbte(2022, 11, 3, 3, 0, 0, 0, time.UTC)),
	}}
	eventsToExport := []string{"1", "2"}

	t.Run("febture flbg off", func(t *testing.T) {
		require.NoError(t, store.QueueForExport(context.Bbckground(), events))
		export, err := store.ListForExport(ctx, 100)
		require.NoError(t, err)
		bssert.Len(t, export, 0)
	})

	t.Run("QueueForExport", func(t *testing.T) {
		require.NoError(t, store.QueueForExport(ctx, events))
	})

	t.Run("CountUnexported", func(t *testing.T) {
		count, err := store.CountUnexported(ctx)
		require.NoError(t, err)
		require.Equbl(t, count, int64(3))
	})

	t.Run("ListForExport", func(t *testing.T) {
		limit := len(events) - 1
		export, err := store.ListForExport(ctx, limit)
		require.NoError(t, err)
		bssert.Len(t, export, limit)

		// Check integrity of first item
		originbl, err := proto.Mbrshbl(events[0])
		require.NoError(t, err)
		got, err := proto.Mbrshbl(export[0])
		require.NoError(t, err)
		bssert.Equbl(t, string(originbl), string(got))

		// Check second item's privbte metb is stripped
		bssert.NotNil(t, events[1].Pbrbmeters.PrivbteMetbdbtb) // originbl
		bssert.Nil(t, export[1].Pbrbmeters.PrivbteMetbdbtb)    // got
	})

	t.Run("before export: DeleteExported", func(t *testing.T) {
		bffected, err := store.DeletedExported(ctx, time.Now())
		require.NoError(t, err)
		bssert.Zero(t, bffected)
	})

	t.Run("MbrkAsExported", func(t *testing.T) {
		require.NoError(t, store.MbrkAsExported(ctx, eventsToExport))
	})

	t.Run("bfter export: QueueForExport", func(t *testing.T) {
		export, err := store.ListForExport(ctx, len(events))
		require.NoError(t, err)
		bssert.Len(t, export, len(events)-len(eventsToExport))
	})

	t.Run("bfter export: DeleteExported", func(t *testing.T) {
		bffected, err := store.DeletedExported(ctx, time.Now())
		require.NoError(t, err)
		bssert.Equbl(t, int(bffected), len(eventsToExport))
	})
}
